Vitest batch 28: test-dashboard-data-protection (17 asserts split)
Split off:
  - test-dashboard-data-protection.js (13 asserts) → Vitest. The five
    CTA-contract scenarios (all-configured / backup-unsupported /
    single-missing direct CTA × 3 / two-missing / all-missing) assert
    against the HTML string from renderDataProtectionCta() — which takes
    an explicit state override, so no capability stub needed — plus the
    section-7 window-export checks.

  - test-dashboard-data-protection-dom.js (4 asserts) → new puppeteer
    file. Section 6: openDataProtectionPicker() attaching a live overlay,
    3-card count, Cancel-button click dismissal.

settings.js is imported alongside context-cards.js so the section-7
window-export check for showSyncSetupModal (exported from settings.js,
not context-cards.js) sees it — puppeteer gets it via main.js.

Vitest: 70 → 71 tests, ~4.44s → ~4.40s.
Puppeteer: 30 → 29 remaining (test-dashboard-data-protection swapped for slim DOM file).
Full run-tests.sh: green.
